Remove cancel and only use finish
We have a cancel action and finish action which was not being used. We
can add a callback action to cancel and I found out I use it to cancel
the action which initiates the action. e.g. focus (init action) / blur
(cancel callback).

This use-case is more for `finish` not `cancel` so I decide to change
`cancel` to `finish`